+++ deep_causality/tests/types/context_types/context_graph/mod.rs
@@ -7,6 +7,8 @@ mod context_tests;
 #[cfg(test)]
 mod contextuable_graph_tests;
 #[cfg(test)]
+mod extendable_context_edge_tests;
+#[cfg(test)]
 mod extendable_context_tests;
 #[cfg(test)]
 mod extendable_contextuable_graph_tests;
